requirements.txt
fastapi
pydantic
datetime
sqlite3
requests
python-dotenv
